renamed auth.ts to authTypes.ts, updated imports
